move: 커스텀 예외 클래스들을 "ladder/domain" -> "ladder/exception"으로 이동함
